search-agent/pyproject.toml
[project]
name = "search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.15",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
]
